Actualizacion

+++ Nuggets/PnlDeleteRepository.cs
@@ -12,6 +12,8 @@ namespace Nuggets
 {
     public partial class PnlDeleteRepository : Form
     {
+        public PnlUser pnlPadre { get; set; }
+        public User user { get; set; }
         public PnlDeleteRepository()
         {
             InitializeComponent();
@@ -19,24 +21,37 @@ namespace Nuggets
 
         private void button2_Click(object sender, EventArgs e)
         {
-            Form1 pnl = new Form1();
-            pnl.Show();
+            pnlPadre.Show();
             this.Hide();
         }
 
         private void button1_Click(object sender, EventArgs e)
         {
             int id = int.Parse(txtUser.Text);
-            int result = DAORepository.delete(id);
 
-            if (result > 0)
+            Repository repo = DAORepository.read(id);
+            
+            if(repo.id_creator == this.user.id)
             {
-                MessageBox.Show("Repositorio eliminado.", "Usuario registrado.", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            }
-            else
+                //Funcion eliminar
+                int result = DAORepository.delete(id);
+                if (result > 0)
+                {
+                    MessageBox.Show("Repositorio eliminado.", "Usuario registrado.", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    pnlPadre.Show();
+                    this.Hide();
+                }
+                else
+                {
+                    MessageBox.Show("No se pudo eliminar el registro", "Fallo al ingresar", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                }
+                //Fin de la funcion eliminar
+            }else
             {
-                MessageBox.Show("No se pudo eliminar el registro", "Fallo al ingresar", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                MessageBox.Show("No tienes permiso de eliminar.", "Fallo al eliminar", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
             }
+
+
         }
     }
 }
